Move internal/pathtools to pathtools
