+++ cmd/move_labels/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//internal/label:go_default_library",
-        "//internal/pathtools:go_default_library",
+        "//pathtools:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
     ],
 )
