+++ cmd/move_labels/move_labels.go
@@ -29,7 +29,7 @@ import (
 	"strings"
 
 	"github.com/bazelbuild/bazel-gazelle/internal/label"
-	"github.com/bazelbuild/bazel-gazelle/internal/pathtools"
+	"github.com/bazelbuild/bazel-gazelle/pathtools"
 	"github.com/bazelbuild/buildtools/build"
 )
