+++ internal/language/go/BUILD.bazel
@@ -51,10 +51,10 @@ go_library(
         "//internal/label:go_default_library",
         "//internal/language:go_default_library",
         "//internal/language/proto:go_default_library",
-        "//internal/pathtools:go_default_library",
         "//internal/repos:go_default_library",
         "//internal/resolve:go_default_library",
         "//internal/rule:go_default_library",
+        "//pathtools:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
     ],
 )
