+++ internal/repos/BUILD.bazel
@@ -12,8 +12,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//internal/label:go_default_library",
-        "//internal/pathtools:go_default_library",
         "//internal/rule:go_default_library",
+        "//pathtools:go_default_library",
         "//vendor/github.com/pelletier/go-toml:go_default_library",
         "//vendor/golang.org/x/tools/go/vcs:go_default_library",
     ],
