+++ internal/repos/remote.go
@@ -25,7 +25,7 @@ import (
 	"sync"
 
 	"github.com/bazelbuild/bazel-gazelle/internal/label"
-	"github.com/bazelbuild/bazel-gazelle/internal/pathtools"
+	"github.com/bazelbuild/bazel-gazelle/pathtools"
 	"golang.org/x/tools/go/vcs"
 )
