+++ pathtools/path.go
