+++ pathtools/path_test.go
